Relocate examples
