+++ examples/gateway.service.ts
@@ -1,7 +1,7 @@
 import type { ServiceBroker } from 'moleculer';
 import { Service } from 'moleculer';
 import ApiService from 'moleculer-web';
-import gatewayMixin from '../gatewayMixin';
+import gatewayMixin from '../src/gatewayMixin';
 
 class GatewayService extends Service {
 	public constructor(broker: ServiceBroker) {
